requirements.txt
streamlit
google-generativeai
requests
streamlit-calendar
pandas
symspellpy
jamo
Pillow
ics